@@ -302,4 +302,4 @@ def create_time_index(
         number = round(hoy / interval)
     if start is None:
         start = f"1/1/{year}"
-    return pd.date_range(start, periods=number + 1, freq=f"{interval}H")
+    return pd.date_range(start, periods=number + 1, freq=f"{interval}h")
