@@ -62,7 +62,7 @@ def equate_variables(model, var1, var2, factor1=1, name=None):
 
     >>> import pandas as pd
     >>> from oemof import solph
-    >>> date_time_index = pd.date_range('1/1/2012', periods=6, freq='H')
+    >>> date_time_index = pd.date_range('1/1/2012', periods=6, freq='h')
     >>> energysystem = solph.EnergySystem(
     ...     timeindex=date_time_index,
     ...     infer_last_interval=False,
