@@ -102,7 +102,7 @@ def generic_integral_limit(om, keyword, flows=None, limit=None):
     --------
     >>> import pandas as pd
     >>> from oemof import solph
-    >>> date_time_index = pd.date_range('1/1/2012', periods=6, freq='H')
+    >>> date_time_index = pd.date_range('1/1/2012', periods=6, freq='h')
     >>> energysystem = solph.EnergySystem(
     ...     timeindex=date_time_index,
     ...     infer_last_interval=False,
