@@ -171,7 +171,7 @@ def additional_investment_flow_limit(model, keyword, limit=None):
     --------
     >>> import pandas as pd
     >>> from oemof import solph
-    >>> date_time_index = pd.date_range('1/1/2020', periods=6, freq='H')
+    >>> date_time_index = pd.date_range('1/1/2020', periods=6, freq='h')
     >>> es = solph.EnergySystem(
     ...     timeindex=date_time_index,
     ...     infer_last_interval=False,
