@@ -1529,7 +1529,7 @@ class TestsMultiPeriodConstraint:
                     nominal_value=100, min=0.32, nonconvex=solph.NonConvex()
                 )
             },
-            coefficients=[-17, 0.9],
+            coefficients={bth: (-17, 0.9)},
         )
         self.energysystem.add(bgas, bth, otrf)
         self.compare_lp_files("offsetconverter_multi_period.lp")
